// lib/types.ts
export interface AutofillExcelRow {
  pod: string
  internalPodId?: string
  podTypeOriginal?: string
  podProgramType?: string
  projectManagers?: string
  clli?: string
  city?: string
  state?: string
  routerType?: string
  router1?: string
  router2?: string
  tenantName?: string
}